Move script arg objects into types module
